Scrape change log from a validation project

Scrape change log from a validation project

Format change log info table for validation report

Initiate a change_log file

vt_scrape_change_log()

vt_kable_change_log(change_log_info, format = vt_render_to())

vt_use_change_log(date = NULL, version = NULL, open = interactive())

Arguments

change_log_info
data.frame as exported from vt_scrape_change_log

format
passed to knitr::kable

date
passed to template

version
version to set in news file

open
whether to open the file after

Value

data.frame with variables version, effective_date, description

a knitr_kable object

path to change log file, used for side effect of creating change_log

Note

Extracts validation version, date, and description from change log items that start with [validation].

Examples

withr::with_tempdir({
 file.create(".here")
 vt_use_validation()

 vt_use_change_log()

 log_data <- vt_scrape_change_log()
 print(log_data)

 vt_kable_change_log(log_data)

})
#> ✔ Adding '!*' to '././validation/.gitignore'
#> Created ./validation  in package structure
#> validation config file created. Add user information through `vt_add_user_to_config()`
#> Change log created in the validation folder.
#>   version effective_date                              description
#> 1     1.0     2021-11-04 Validation release notes for version 1.0
#> <table>
#>  <thead>
#>   <tr>
#>    <th style="text-align:left;"> Version </th>
#>    <th style="text-align:left;"> Effective Date </th>
#>    <th style="text-align:left;"> Activity Description </th>
#>   </tr>
#>  </thead>
#> <tbody>
#>   <tr>
#>    <td style="text-align:left;border-left:1px solid;"> 1.0 </td>
#>    <td style="text-align:left;"> 2021-11-04 </td>
#>    <td style="text-align:left;width: 30em; border-right:1px solid;"> Validation release notes for version 1.0 </td>
#>   </tr>
#> </tbody>
#> </table>