Validate a package

vt_validate_source runs the validation on the current source, temporarily installing the source to properly evaluate the report. vt_validate_build() runs the same step, then compiles a bundle that includes th the validation report and any other contents that are required for validation. Finally, vt_validate_installed_package run rerun the validation report for packages that were built and then installed using thevt_validate_build().

Usage

vt_validate_source(src = ".", open = interactive())

vt_validate_build(src = ".", ...)

vt_validate_install(
  src = ".",
  ...,
  install_verbose = TRUE,
  install_tests = TRUE,
  reload = TRUE
)

vt_validate_installed_package(
  package,
  output_directory = ".",
  open = interactive()
)

vt_validate_report(version, open = interactive())

Arguments

src

location of the source code. Assumed to be the same location as "pkg"

open

should the validation report be opened after it is built?

...

Additional argument passed to devtools::build()

install_verbose

should the installation be verbose?

install_tests

should the installation include installation of package-specific tests (if any)?

reload

Should package be reloaded after install? defaults to TRUE

package

installed package name

output_directory

Location of directory to output validation report

version

version of validation report to output. If missing, it tries to use the change_log.md, if that is missing then looks at the package version if the validation package is of an R package.

Value

path to either the validation report or the bundled package