Retrieve roxygen tags as a data.frame from requirements, test cases, test code and functions

Looks for roxygen2 function documentation in /R for author details. Assumes that author and date are tagged via custom sections @section Last updated by: and @section Last updated date:, respectively. To exclude a roxygen block from this scraping, omit these section names.

If using a dummy documentation file, looks for @name to capture function name, otherwise uses the actual function call.

Exported or internal status does not affect scraping.

vt_scrape_tags_from(
  type,
  tags = c("editor", "editDate"),
  src = ".",
  ref = vt_path()
)

Arguments

type
type of scraping to be done. one of "requirements","test_cases","test_code","functions". to call functions. working directory must be an R package, or path identified in src must be an R package.

tags
which tags to keep. defaults to editor and editDate

src
path to package source. defaults to the current directory.

ref
reference path to where validation documentation lives. defaults to vt_path()

Note

At this time, this function does not retrieve documentation captured for functions dispatched within an R6 class. Tags at the class level documentation are retrieved.

Last Updated by

Ellis Hughes

Last updated date

2021-03-05