Create a validation structure

Creates a structure for validation artifacts. Validation items are stored in inst/validation

vt_use_validation(pkg = ".", working_dir, ...)

vt_create_package(
  pkg = ".",
  ...,
  fields = list(),
  rstudio = rstudioapi::isAvailable(),
  roxygen = TRUE,
  check_name = TRUE,
  open = rlang::is_interactive()
)

Arguments

pkg
Top level directory of a package

working_dir
validation working directory of the project. Defaults to

...
Additional argument passed to vt_use_config()

fields
A named list of fields to add to DESCRIPTION, potentially overriding default values. See use_description() for how you can set personalized defaults using package options

rstudio
If TRUE, calls use_rstudio() to make the new package or project into an RStudio Project. If FALSE and a non-package project, a sentinel .here file is placed so that the directory can be recognized as a project by the here or rprojroot packages.

roxygen
Do you plan to use roxygen2 to document your package?

check_name
Whether to check if the name is valid for CRAN and throw an error if not

open
If TRUE, activates the new project:

If RStudio desktop, the package is opened in a new session.

If on RStudio server, the current RStudio project is activated.

Otherwise, the working directory and active project is changed.