Use dynamic file paths in a validation.

vt_path() allows access of files relative to the working directory,which is identified by the config file. It is also required to be used in the validation report for cases where validation of installed packages is intended as it will shift access to the correct location for the installed package for access.

Usage

vt_path(...)

vt_find_config()

Arguments

...

[character]
Path components below the validation folder, can be empty. Each argument should be a string containing one or more path components separated by a forward slash "/".

Details

vt_find_config() locates the config file in the working directory, and returns the full path to it.

Examples

withr::with_tempdir({callr::r(function(){

 valtools::vt_use_validation()

 valtools::vt_path()
 valtools::vt_path("some", "reqs", "req01.md")
 valtools::vt_path("some/reqs/req01.md")

 valtools::vt_find_config()

})})
#> [1] "/tmp/RtmpNdZi6T/file18037914ba04/validation/validation.yml"